+++ com/posix_string.c
@@ -18,7 +18,7 @@ char *posix__trim(char *src) {
         return NULL;
     }
 
-    cb = strlen(src);
+    cb = (int)strlen(src);
     p = src;
     target = src;
 
@@ -30,7 +30,7 @@ char *posix__trim(char *src) {
         posix__strncpy(target, cb, p, (cb - (p - src)));
     }
 
-    cb = strlen(target);
+    cb = (int)strlen(target);
     p = &target[cb - 1];
     while ((*p == '\r' || *p == '\n' || *p == ' ') && (p >= target)) p--;
